add: iso639-1 language codes to resources
A resource is always provided in a specific language. For machine translations etc, languageTags are required.

Furthermore, the files are unbundled for better comprehension what is exported (index.ts) and what not.
